refactor and add logger

+++ index.js
@@ -1,6 +1,10 @@
 import wwebjs from "whatsapp-web.js";
 import qrcode from "qrcode-terminal";
 import { db } from "./src/utils/db.js";
+import {
+  listenedGroupsLogger,
+  generalGroupsLogger,
+} from "./src/utils/logger.js";
 // import { config } from "./config.js";
 
 const { Client, LocalAuth, MessageMedia } = wwebjs;
@@ -47,48 +51,68 @@ client.on("ready", () => {
 
 client.on("message_create", async (msg) => {
   let chat = await msg.getChat();
+  let contactInfo = await msg.getContact();
+  let user = contactInfo.name;
 
   // spew out incoming message to the terminal
   // console.log("Received message:", msg.body);
+  generalGroupsLogger.info(`${chat.name} | ${user} | ${msg.body}`);
 
   if (!chat.id) {
     console.log("Chat ID is undefined");
     return;
   }
 
-  const isListening = db
-    .prepare("SELECT id FROM listened_groups WHERE id = ?")
-    .get(chat.id._serialized);
-  console.log("Chat ID:", chat.id);
-  console.log("Is listening:", isListening);
-
-  if (!isListening) return;
-
-  console.log("Received message:", msg.body);
-
-  if (msg.body === "siapa?") {
-    msg.reply(msg.author || msg.from);
-  }
-  if (msg.body === ".test") {
-    msg.react("😼");
-  }
-
   if (msg.body === "/flagged") {
     if (!chat.isGroup) return;
+    if (!chat.fromMe) return;
 
     db.prepare(
       "INSERT OR IGNORE INTO listened_groups (id, name) VALUES (?, ?)",
     ).run(chat.id._serialized, chat.name);
 
     msg.react("📍");
+    listenedGroupsLogger.info(
+      `Started listening to group: ${chat.name} (${chat.id._serialized})`,
+    );
+    return;
   }
 
   if (msg.body === "/unflagged") {
     if (!chat.isGroup) return;
+    if (!chat.fromMe) return;
+
     db.prepare("DELETE FROM listened_groups WHERE id = ?").run(
       chat.id._serialized,
     );
     msg.react("🤐");
+    listenedGroupsLogger.info(
+      `Stopped listening to group: ${chat.name} (${chat.id._serialized})`,
+    );
+    return;
+  }
+
+  if (msg.body === "/get contact") {
+    let contact = await msg.getContact();
+    msg.reply(JSON.stringify(contact, null, 2));
+  }
+
+  const isListening = db
+    .prepare("SELECT id FROM listened_groups WHERE id = ?")
+    .get(chat.id._serialized);
+
+  // sacred line to prevent the bot from responding to messages in groups that are not flagged as listened
+  if (!isListening) return;
+
+  listenedGroupsLogger.info(`Message from ${chat.name}: ${msg.body}`);
+
+  if (msg.body === "siapa?") {
+    msg.reply(user);
+    listenedGroupsLogger.info(`Replied to ${chat.name} with sender info.`);
+  }
+  if (msg.body === ".test") {
+    msg.react("😼");
+    listenedGroupsLogger.info(`Reacted to message from ${chat.name}.`);
   }
 
   if (msg.body === "/get chat") {
